Move MultiObjectDeleteException to impl
